Timer state moved to UserHome component
